THE 1.8G UPDATE (Kakudai flowers)
-THE 1.8G UPDATE (AKA 'Kakudai flowers') (NONE):

note: from this update, only the updates that need changes of WBSE compared
to the version before it  would be added as a new version in the W series, and only the update that need changes for their AMM compared to the version before it would be added as a new version in the A series.

this update is an improvement to the interface , firstly we minimized the required width and height that it become even less than Microsoft edge's minimum width and height, secondly we improved the homepage and incognito a lot like a lot we made them have different cases and made them working much better with smaller dimensions esp. width , finally we done some other GUI changes like removing the spaces between the 'new tab' button and 'extensions' button and even removed the space between the 'refresh' button and URL and name bars plus we minimized the minimum bars width from 100 pixels to 20 pixels.